move pipeline agnostic functions to base module
